+++ packages/bundler/src/cli/commands/run/server.ts
@@ -4,9 +4,9 @@ import { BundlerServer } from "../../../server";
 import { getRedis } from "../../utils";
 
 const runServer = new Command("server")
-  .summary("Run bundler server")
+  .summary("run bundler server")
   .description(
-    "Must supply .env file with REDIS_URL and RPC_URL. Must supply wallet contract address and port as options."
+    "must supply .env file with REDIS_URL and RPC_URL. must supply wallet contract address and port as options."
   )
   .requiredOption("--wallet-address <string>", "wallet contract address")
   .requiredOption("--port <number>", "server port", parseInt)
@@ -15,12 +15,12 @@ const runServer = new Command("server")
 
     const rpcUrl = process.env.RPC_URL;
     if (!rpcUrl) {
-      throw new Error("Missing RPC_URL");
+      throw new Error("missing RPC_URL");
     }
     const provider = new ethers.providers.JsonRpcProvider(rpcUrl);
 
     const server = new BundlerServer(walletAddress, provider, getRedis());
-    server.run(port);
+    server.start(port);
   });
 
 export default runServer;