+++ packages/e2e-tests/src/anvil.ts
@@ -1,17 +1,23 @@
-import { RunCommandDetachedOpts, runCommandDetached, sleep } from "./utils";
+import { ethers } from "ethers";
+import {
+  ResetFn,
+  RunCommandDetachedOpts,
+  runCommandBackground,
+  sleep,
+} from "./utils";
+import { assertOrErr } from "@nocturne-xyz/deploy/dist/src/utils";
 
 export interface AnvilNetworkConfig {
   blockTimeSecs?: number;
   gasPrice?: bigint;
 }
 
-export async function startAnvil(
-  config: AnvilNetworkConfig
-): Promise<() => Promise<void>> {
+// returns snapshotId of empty chain state
+export async function startAnvil(config: AnvilNetworkConfig): Promise<ResetFn> {
   const { blockTimeSecs, gasPrice } = config ?? {};
 
   const cmd = "anvil";
-  const args = [];
+  const args = ["--host", "0.0.0.0"];
 
   if (blockTimeSecs) {
     args.push("--block-time", blockTimeSecs.toString());
@@ -26,9 +32,21 @@ export async function startAnvil(
     onError: console.error,
   };
 
-  const stop = runCommandDetached(cmd, args, cmdOpts);
-  await sleep(5_000);
+  runCommandBackground(cmd, args, cmdOpts);
+  await sleep(1_000);
+
+  const provider = new ethers.providers.JsonRpcProvider("http://0.0.0.0:8545");
+  // get snapshot with empty chain state
+  let snapshotId = await provider.send("evm_snapshot", []);
   return async () => {
-    stop();
+    const provider = new ethers.providers.JsonRpcProvider(
+      "http://0.0.0.0:8545"
+    );
+
+    const success = await provider.send("evm_revert", [snapshotId]);
+    assertOrErr(success, "failed to revert to snapshot");
+
+    await sleep(1_000);
+    snapshotId = await provider.send("evm_snapshot", []);
   };
 }