+++ scripts/run_site_dev.sh
@@ -22,6 +22,11 @@ yarn start &
 SITE_PID=$!
 popd
 
+# wait a bit for site to finish building
+# this is to avoid gatsby throwing a fit because we have stuff running on the
+# same ports its dumb graphql server wants to use
+sleep 20
+
 # start the snap
 pushd packages/snap
 echo "starting snap..."
@@ -32,7 +37,7 @@ popd
 # start anvil
 pushd packages/e2e-tests
 echo "starting anvil..."
-anvil &> "$LOG_DIR/anvil" &
+anvil --block-time 1 --host 0.0.0.0 &> "$LOG_DIR/anvil" &
 ANVIL_PID=$!
 
 sleep 1
@@ -60,14 +65,10 @@ SUBTREE_UPDATER_TX_SIGNER_KEY="0x8b3a350cf5c34c9194ca85829a2df0ec3153be0318b5e2d
 # eth address: 0x976EA74026E726554dB657fA54763abd0C3a0aa9
 SCREENER_TX_SIGNER_KEY="0x92db14e403b83dfe3df233f83dfa3a0d7096f21ca9b0d6d6b8d88b2b4ec1564e"
 
-# deposit
-echo "Running deposit funds script..."
-yarn hh-node-deposit &> "$LOG_DIR/hh-node-deposit" || { echo 'hh-node-deposit failed' ; exit 1; }
-
 # read config variables from logs
 read DEPOSIT_MANAGER_CONTRACT_ADDRESS < <(sed -nr 's/^DepositManager address: (0x[a-fA-F0-9]{40})$/\1/p' $LOG_DIR/anvil-deposit)
 read WALLET_CONTRACT_ADDRESS < <(sed -nr 's/^Wallet address: (0x[a-fA-F0-9]{40})$/\1/p' $LOG_DIR/anvil-deposit)
-read VAULT_CONTRACT_ADDRESS < <(sed -nr 's/^Vault address: (0x[a-fA-F0-9]{40})$/\1/p' $LOG_DIR/anvil-deposit)
+read HANDLER_CONTRACT_ADDRESS< <(sed -nr 's/^Handler address: (0x[a-fA-F0-9]{40})$/\1/p' $LOG_DIR/anvil-deposit)
 read TOKEN_CONTRACT_ADDR1 < <(sed -nr 's/^Token 1 deployed at: (0x[a-fA-F0-9]{40})$/\1/p' $LOG_DIR/anvil-deposit)
 read TOKEN_CONTRACT_ADDR2 < <(sed -nr 's/^Token 2 deployed at: (0x[a-fA-F0-9]{40})$/\1/p' $LOG_DIR/anvil-deposit)
 read GAS_TOKEN_CONTRACT_ADDR < <(sed -nr 's/^Gas token deployed at: (0x[a-fA-F0-9]{40})$/\1/p' $LOG_DIR/anvil-deposit)